DRYer implementation of new_export_instance
